requirements.txt
fastapi
uvicorn
sqlalchemy
pydantic
python-dotenv
MetaTrader5
